Fix incorrect portrait/landscape on N900.
These values were totally different from what meego-qml-launcher uses (as
determined from the Qt Mobility framework). Qt Mobility gives the correct
orientation, which is then translated into an orientation passed to QML, but it
appears the original author of this QML was using a device which had some
schitzophrenia as to which way was 'up'.

The 'fix' to do portrait/landscape using the incorrect values is not a fix,
however, because third party applications will still have misleading Qt Mobility
readings, so either the device needs to somehow alter the readings returned from
mobility, or the screen display should somehow be rotated.

Signed-off-by: Robin Burchell <viroteck@viroteck.net>