[browserlistmodel] Use db connection name to avoid conflict
This fix made it into standard components long ago, but that's not
being used. If two different people in the same process (like this
and QMF) both use a default database connection name, one trashes
the other. Since we can't easily fix QMF, fix this one and we have
to make sure we don't add this mistake again elsewhere.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>